requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
build==1.2.2
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cleo==2.1.0
constantly==23.10.4
crashtest==0.4.1
cryptography==43.0.1
daphne==4.1.2
distlib==0.3.8
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.0
future==1.0.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.3.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyproject_hooks==1.1.0
PyYAML==6.0.2
rapidfuzz==3.9.7
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.20.0
SecretStorage==3.3.3
serializers==0.2.4
service-identity==24.1.0
shellingham==1.5.4
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
trove-classifiers==2024.9.12
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.4
zipp==3.20.2
zope.interface==7.0.3
